Relocate the opaque session directory into wow-world
Issue #138 moves the already-opaque connected-session directory out of raw
networking. Identifying live session incarnations and resolving their control
endpoints is session coordination, not socket transport, so `PlayerRegistry`,
its private `PlayerRegistryEntry` storage, `PlayerBroadcastInfo`, the
generation-aware `PlayerRegistration`/`PlayerControlAddress` addressing types,
the send-outcome enums, every named projection and the directory test now live
in `wow_world::session::directory`.

This is a mechanical relocation. No gameplay, packet bytes, observable
ordering, clock, persistence or Player ownership changes, and the directory
test moves verbatim.

`wow-network` keeps everything that is genuinely transport or mailbox:
`SessionCommand` and its 42 payload types, `SharedClientVisibleGuidsLikeCpp`,
the durable creature-runtime and loot-money rails, the physical byte channels,
`SocketWriteFenceLikeCpp` and `InstanceLink`. The pump and mailbox boundary
stay with #140, durable loot persistence with #189, and `PlayerBroadcastInfo`
retirement with #252. The directory's only remaining edge into networking is
the existing downward `wow-world -> wow-network` dependency, so no cycle is
created and no service locator, event bus or second mutable mirror is added.

Obsolete surface is removed rather than re-exported: the 21 crate-root
`wow_network` re-exports of directory types are gone, and the now-unused
`wow-network` `test-fixtures` feature plus the dev-dependencies that enabled it
are deleted. The four feature-gated fixture helpers follow the directory into
`wow-world` behind its own `test-fixtures` feature, which `world-server` enables
for its tests.

Guards follow the owner instead of being weakened. `session-ownership-check`
now scans `crate::session::directory` for the session contract surfaces, so
`PlayerBroadcastInfo` is still inventoried; its "network type" concept is
renamed to the accurate "session contract type". The anti-glob registry guard
gains the relocated owner, with a new adversarial test proving that
`use wow_world::session::directory::*;` and `use crate::session::directory::*;`
fail closed while an unrelated session submodule glob stays allowed.

Reviewed baseline delta: `PlayerBroadcastInfo` keeps exactly 80 identical
fields and only its module moves; one derive input remaps to the new module;
one signature shortens to the imported `PlayerVisibilityCreateSnapshot`; the
573 direct-registry rows fall to 572 because the removed `wow_network`
re-export disappears, with every other row identical modulo import-list text;
the 71 bridge rows keep identical identities. `WorldSession` fields (738),
impl items (3341), `SessionCommand` variants (37) and payload types (42) are
unchanged, and `PlayerRegistry` still exposes zero generic
get/get_mut/iter/insert/remove/clear operations. The Session logical hotspot
ceiling rises by exactly the 1,773 production and 134 test relocated lines,
documented in the ownership evidence section.

Validation: cargo check --workspace --all-targets (clean, no new warnings),
cargo test -p wow-world --lib session::directory:: and registry (34 tests,
including generation, replacement, stale-unregister and control-address
identity), cargo test -p wow-network --lib (125), world-server delivery and
directory-generation tests (8), the handler-contract-check registry_access
tests (7), session-ownership-check check --syntax-only, check_architecture.py
check and self-test, cargo fmt --all -- --check and git diff --check.

Closes #138
