+++ crates/wow-network/Cargo.toml
@@ -5,9 +5,6 @@ edition.workspace = true
 rust-version.workspace = true
 license.workspace = true
 
-[features]
-test-fixtures = []
-
 [dependencies]
 tokio = { workspace = true }
 bytes = { workspace = true }
